+++ src/TourOfHeroes.Web/Pages/Heroes/Containers/HeroDetails.razor.cs
@@ -1,7 +1,7 @@
 using BlazorState;
 using Microsoft.AspNetCore.Components;
 using TourOfHeroes.Web.Pages.Heroes.Models;
-using TourOfHeroes.Web.Pages.Heroes.Feature;
+using TourOfHeroes.Web.Pages.Heroes.State;
 
 namespace TourOfHeroes.Web.Pages.Heroes.Containers
 {
