+++ src/TourOfHeroes.Web/Pages/Heroes/Containers/Heroes.razor.cs
@@ -2,7 +2,7 @@ using System.Linq;
 using BlazorState;
 using Microsoft.AspNetCore.Components;
 using TourOfHeroes.Web.Pages.Heroes.Models;
-using TourOfHeroes.Web.Pages.Heroes.Feature;
+using TourOfHeroes.Web.Pages.Heroes.State;
 
 namespace TourOfHeroes.Web.Pages.Heroes.Containers
 {
