+++ src/TourOfHeroes.Web/Pages/Heroes/State/Create/CreateAction.cs
@@ -1,6 +1,6 @@
 using BlazorState;
 
-namespace TourOfHeroes.Web.Pages.Heroes.Feature
+namespace TourOfHeroes.Web.Pages.Heroes.State
 {
     public partial class HeroesState 
     {
