+++ src/TourOfHeroes.Web/Pages/Heroes/State/Create/HandleCreate.cs
@@ -6,7 +6,7 @@ using MediatR;
 using TourOfHeroes.Web.Shared;
 using TourOfHeroes.Web.Pages.Heroes.Models;
 
-namespace TourOfHeroes.Web.Pages.Heroes.Feature
+namespace TourOfHeroes.Web.Pages.Heroes.State
 {
     public partial class HeroesState
     {
