+++ src/TourOfHeroes.Web/Pages/Heroes/State/Destroy/DestroyAction.cs
@@ -1,14 +1,14 @@
 using BlazorState;
 
-namespace TourOfHeroes.Web.Pages.Heroes.Feature
+namespace TourOfHeroes.Web.Pages.Heroes.State
 {
     public partial class HeroesState 
     {
-        public class RetrieveOneAction : IAction
+        public class DestroyAction: IAction
         {
             public int Id { get; private set; }
 
-            public RetrieveOneAction(int id)
+            public DestroyAction(int id)
             {
                 Id = id;
             }
