+++ src/TourOfHeroes.Web/Pages/Heroes/State/HeroesState.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using BlazorState;
 using TourOfHeroes.Web.Pages.Heroes.Models;
 
-namespace TourOfHeroes.Web.Pages.Heroes.Feature
+namespace TourOfHeroes.Web.Pages.Heroes.State
 {
     public partial class HeroesState : State<HeroesState>
     {
