+++ src/TourOfHeroes.Web/Pages/Heroes/State/RetrieveOne/HandleRetrieveOne.cs
@@ -7,7 +7,7 @@ using MediatR;
 using TourOfHeroes.Web.Pages.Heroes.Models;
 using TourOfHeroes.Web.Shared;
 
-namespace TourOfHeroes.Web.Pages.Heroes.Feature
+namespace TourOfHeroes.Web.Pages.Heroes.State
 {
     public partial class HeroesState
     {
