+++ src/TourOfHeroes.Web/Pages/Heroes/State/RetrieveOne/RetrieveOneAction.cs
@@ -1,14 +1,14 @@
 using BlazorState;
 
-namespace TourOfHeroes.Web.Pages.Heroes.Feature
+namespace TourOfHeroes.Web.Pages.Heroes.State
 {
     public partial class HeroesState 
     {
-        public class DestroyAction: IAction
+        public class RetrieveOneAction : IAction
         {
             public int Id { get; private set; }
 
-            public DestroyAction(int id)
+            public RetrieveOneAction(int id)
             {
                 Id = id;
             }
