+++ src/TourOfHeroes.Web/Startup.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using TourOfHeroes.Web.Pages.Heroes.Feature;
+using TourOfHeroes.Web.Pages.Heroes.State;
 using TourOfHeroes.Web.Shared;
 
 namespace TourOfHeroes.Web
